empapp-py-client/requirements.txt
sseclient-py
urllib3